+++ src/mixins/ViewsMixin.js
@@ -19,6 +19,10 @@
  * along with this program. If not, see <http://www.gnu.org/licenses/>.
  */
 
+import { generateUrl } from '@nextcloud/router'
+import { showError } from '@nextcloud/dialogs'
+import axios from '@nextcloud/axios'
+
 export default {
 	props: {
 		hash: {
@@ -28,6 +32,23 @@ export default {
 		form: {
 			type: Object,
 			required: true,
+		},
+	},
+
+	methods: {
+		async saveFormProperty(key) {
+			try {
+				// TODO: add loading status feedback ?
+				await axios.post(generateUrl('/apps/forms/api/v1/form/update/'), {
+					id: this.form.id,
+					keyValuePairs: {
+						[key]: this.form[key],
+					},
+				})
+			} catch (error) {
+				showError(t('forms', 'Error while saving form'))
+				console.error(error)
+			}
 		},
 	},
 }
